+++ include/mkx.h
@@ -64,7 +64,6 @@ extern int	mkintfname(char *,cchar *,cchar *,cchar *) noex ;
 extern int	mktagfname(char *,cchar *,cchar *,int) noex ;
 extern int	mkmid(char *,int,cchar *,cchar *,pid_t,int) noex ;
 extern int	mkonefrom(char *,int,cchar *,int) noex ;
-extern int	mkbestfrom(char *,int,cchar *,int) noex ;
 extern int	mkmailname(char *,int,cchar *,int) noex ;
 extern int	mkvarpath(char *,cchar *,int) noex ;
 extern int	mkexpandpath(char *,cchar *,int) noex ;
