+++ src/libpcs/pcsmailcheck.c
@@ -152,8 +152,8 @@ extern int	sfsubstance(const char *,int,const char **) ;
 extern int	matkeystr(const char **,char *,int) ;
 extern int	vstrkeycmp(const char **,const char **) ;
 extern int	pathclean(char *,const char *,int) ;
-extern int	mkbestfrom(char *,int,const char *,int) ;
-extern int	mkdisphdr(char *,int,cchar *,int) ;
+extern int	mkaddrfrom(char *,int,const char *,int) ;
+extern int	mkaddrdisp(char *,int,cchar *,int) ;
 extern int	getusername(char *,int,uid_t) ;
 extern int	getuid_name(cchar *,int) ;
 extern int	mailbox_getfrom(MAILBOX *,char *,int,cchar *,int) ;
@@ -599,7 +599,7 @@ static int subinfo_cvtfrom(SUBINFO *sip)
 		    if ((rs = uc_malloc((dlen+1),&dbuf)) >= 0) {
 			if (wl > sip->rlen) wl = sip->rlen ;
 		        if ((rs = snwcpywidehdr(dbuf,dlen,wbuf,wl)) >= 0) {
-			    rs = mkdisphdr(sip->rbuf,sip->rlen,dbuf,rs) ;
+			    rs = mkaddrdisp(sip->rbuf,sip->rlen,dbuf,rs) ;
 		            len = rs ;
 		        }
 			rs1 = uc_free(dbuf) ;
